refactor(plugins): split plugin-profiles.js into a bounded module directory
scripts/lib/plugin-profiles.js was 1390 lines with a 188-line
resolvePluginProfilePlan and a 139-line generateProfilePlugin, against a
repository guideline of files under 800 lines and functions under 50.

The library now lives in scripts/lib/plugin-profiles/ as constants,
fs-utils, frontmatter, require-graph, plan, ledger, carrier, and
marketplace, with index.js re-exporting the public surface unchanged.
scripts/lib/plugin-profiles.js becomes a one-line re-export so existing
callers and the carrier dependency closure keep the same require path.

Extracted from resolvePluginProfilePlan: expandSurface (module and
component expansion), resolveFullClosure and coverDependencies (the
require-graph walk), assemblePlan, and collectBlockers, which now holds
every refusal in one place instead of inline in previewProfilePlugin.

Extracted from generateProfilePlugin: buildStagingTree,
verifyStagedCarrier, buildReceipt, writeReceipt, swapIntoPlace, and
stageVerifyAndSwap.

Behaviour is unchanged: the hook decision is still derived after the
dependency closure is folded in, so a dependency that is itself a held
hook path still makes the decision pending.

scripts/ci/check-module-size.js reports file and function lengths and can
gate on them. Longest file is now carrier.js at 626 lines; longest
function is previewProfilePlugin at 44.

manifests/install-modules.json needs no change: commands-core lists the
CLI entry point (scripts/plugin-profiles.js) and hooks-runtime lists
scripts/lib as a directory, both of which still cover the new layout.

Closes CodeRabbit finding: scripts/lib/plugin-profiles.js is ~1338 lines,
well past the 800-line guideline, with functions well past 50 lines.
